Create several sections for kernel-specific settings For each kernel separately and using the kernels available in GitHub, find them from GitHub and add them. Just for example, for the xray kernel, use this https://github.com/MHSanaei/3x-ui After this, put it all in the kernels section at the bottom of the users
In the users section, put a switch to turn off the user and not use it and a status display. Change the volume status section from a circle to a bar that is filling up in green and finally turns red, i.e. the last 1 gig is red and the last 5 gig is yellow. The protocols of each are different, i.e. xray itself has vlss, ws and vmss, so all of these must be brought into the user by specifying the kernel, i.e. when it is specified, xray brings its protocols.
There must be a link or qr download section. For supported accounts, for example, for open vpn, the file must be downloaded. Get this information from GitHub